+++ src/excel/excel_functions.rb
@@ -150,3 +150,5 @@ require_relative 'excel_functions/rate'
 require_relative 'excel_functions/ceiling'
 
 require_relative 'excel_functions/mround'
+
+require_relative 'excel_functions/exact'